+++ gamemodes/terrortown/gamemode/shared/sh_init.lua
@@ -522,6 +522,7 @@ include("ttt2/libraries/thermalvision.lua")
 include("ttt2/libraries/events.lua")
 include("ttt2/libraries/none.lua")
 include("ttt2/libraries/targetid.lua")
+include("ttt2/libraries/migrations.lua")
 
 -- include ttt required files
 ttt_include("sh_decal")